Removed redundant casts

+++ hoot-core/src/main/cpp/hoot/core/algorithms/changeset/ChangesetReplacementCreatorAbstract.cpp
@@ -568,7 +568,7 @@ OsmMapPtr ChangesetReplacementCreatorAbstract::_getCookieCutMap(
   // map and replace them with small square shaped polys...from that we can generate the alpha
   // shape.
   const int cutterMapToUseSize = cutterMapToUse->getNodeCount();
-  if ((int)cutterMapToUseSize < 3 && MapUtils::mapIsPointsOnly(cutterMapToUse))
+  if (cutterMapToUseSize < 3 && MapUtils::mapIsPointsOnly(cutterMapToUse))
   {
     LOG_DEBUG("Creating a cutter shape map transformation for point map...");
     // Make a copy here since we're making destructive changes to the geometry here for alpha shape
